+++ example/lib/face_detection_page.dart
@@ -8,16 +8,15 @@ class FaceDetectionPage extends StatefulWidget {
 }
 
 class _FaceDetectionPageState extends State<FaceDetectionPage> {
-  ARKitController arkitController;
-  ARKitNode node;
-  String anchorId;
+  late ARKitController arkitController;
+  ARKitNode? node;
 
-  ARKitNode leftEye;
-  ARKitNode rightEye;
+  ARKitNode? leftEye;
+  ARKitNode? rightEye;
 
   @override
   void dispose() {
-    arkitController?.dispose();
+    arkitController.dispose();
     super.dispose();
   }
 
@@ -43,17 +42,15 @@ class _FaceDetectionPageState extends State<FaceDetectionPage> {
       return;
     }
     final material = ARKitMaterial(fillMode: ARKitFillMode.lines);
-    final ARKitFaceAnchor faceAnchor = anchor;
-    faceAnchor.geometry.materials.value = [material];
+    anchor.geometry.materials.value = [material];
 
-    anchorId = anchor.identifier;
-    node = ARKitNode(geometry: faceAnchor.geometry);
-    arkitController.add(node, parentNodeName: anchor.nodeName);
+    node = ARKitNode(geometry: anchor.geometry);
+    arkitController.add(node!, parentNodeName: anchor.nodeName);
 
-    leftEye = _createEye(faceAnchor.leftEyeTransform);
-    arkitController.add(leftEye, parentNodeName: anchor.nodeName);
-    rightEye = _createEye(faceAnchor.rightEyeTransform);
-    arkitController.add(rightEye, parentNodeName: anchor.nodeName);
+    leftEye = _createEye(anchor.leftEyeTransform);
+    arkitController.add(leftEye!, parentNodeName: anchor.nodeName);
+    rightEye = _createEye(anchor.rightEyeTransform);
+    arkitController.add(rightEye!, parentNodeName: anchor.nodeName);
   }
 
   ARKitNode _createEye(Matrix4 transform) {
@@ -63,7 +60,7 @@ class _FaceDetectionPageState extends State<FaceDetectionPage> {
       transform.getColumn(3).z,
     );
     final material = ARKitMaterial(
-      diffuse: ARKitMaterialProperty(color: Colors.yellow),
+      diffuse: ARKitMaterialProperty.color(Colors.yellow),
     );
     final sphere = ARKitBox(
         materials: [material], width: 0.03, height: 0.03, length: 0.03);
@@ -72,13 +69,13 @@ class _FaceDetectionPageState extends State<FaceDetectionPage> {
   }
 
   void _handleUpdateAnchor(ARKitAnchor anchor) {
-    if (anchor is ARKitFaceAnchor) {
-      final ARKitFaceAnchor faceAnchor = anchor;
-      arkitController.updateFaceGeometry(node, anchor.identifier);
-      _updateEye(leftEye, faceAnchor.leftEyeTransform,
-          faceAnchor.blendShapes['eyeBlink_L']);
-      _updateEye(rightEye, faceAnchor.rightEyeTransform,
-          faceAnchor.blendShapes['eyeBlink_R']);
+    if (anchor is ARKitFaceAnchor && mounted) {
+      final faceAnchor = anchor;
+      arkitController.updateFaceGeometry(node!, anchor.identifier);
+      _updateEye(leftEye!, faceAnchor.leftEyeTransform,
+          faceAnchor.blendShapes['eyeBlink_L'] ?? 0);
+      _updateEye(rightEye!, faceAnchor.rightEyeTransform,
+          faceAnchor.blendShapes['eyeBlink_R'] ?? 0);
     }
   }