+++ example/lib/real_time_updates.dart
@@ -1,6 +1,7 @@
 import 'package:arkit_plugin/arkit_plugin.dart';
 import 'package:flutter/material.dart';
 import 'package:vector_math/vector_math_64.dart' as vector;
+import 'package:collection/collection.dart';
 
 class RealTimeUpdatesPage extends StatefulWidget {
   @override
@@ -8,13 +9,14 @@ class RealTimeUpdatesPage extends StatefulWidget {
 }
 
 class _RealTimeUpdatesPageState extends State<RealTimeUpdatesPage> {
-  ARKitController arkitController;
-  ARKitNode movingNode;
+  late ARKitController arkitController;
+  ARKitNode? movingNode;
   bool busy = false;
 
   @override
   void dispose() {
-    arkitController?.dispose();
+    arkitController.updateAtTime = null;
+    arkitController.dispose();
     super.dispose();
   }
 
@@ -33,8 +35,8 @@ class _RealTimeUpdatesPageState extends State<RealTimeUpdatesPage> {
   }
 
   void _onARKitViewCreated(ARKitController arkitController) {
-    final ARKitMaterial material = ARKitMaterial(
-      diffuse: ARKitMaterialProperty(color: Colors.white),
+    final material = ARKitMaterial(
+      diffuse: ARKitMaterialProperty.color(Colors.white),
     );
 
     final sphere = ARKitSphere(
@@ -53,9 +55,9 @@ class _RealTimeUpdatesPageState extends State<RealTimeUpdatesPage> {
         busy = true;
         this.arkitController.performHitTest(x: 0.25, y: 0.75).then((results) {
           if (results.isNotEmpty) {
-            final point = results.firstWhere(
-                (o) => o.type == ARKitHitTestResultType.featurePoint,
-                orElse: () => null);
+            final point = results.firstWhereOrNull(
+              (o) => o.type == ARKitHitTestResultType.featurePoint,
+            );
             if (point == null) {
               return;
             }
@@ -64,11 +66,11 @@ class _RealTimeUpdatesPageState extends State<RealTimeUpdatesPage> {
               point.worldTransform.getColumn(3).y,
               point.worldTransform.getColumn(3).z,
             );
-            final ARKitNode newNode = ARKitNode(
+            final newNode = ARKitNode(
               geometry: sphere,
               position: position,
             );
-            this.arkitController.remove(movingNode.name);
+            this.arkitController.remove(movingNode!.name);
             movingNode = null;
             this.arkitController.add(newNode);
             movingNode = newNode;
@@ -78,6 +80,6 @@ class _RealTimeUpdatesPageState extends State<RealTimeUpdatesPage> {
       }
     };
 
-    this.arkitController.add(movingNode);
+    this.arkitController.add(movingNode!);
   }
 }